manageVideosRow component created
