+++ src/components/Dashboard/ManageVideos.js
@@ -1,6 +1,6 @@
 import { useQuery } from '@tanstack/react-query';
 import React from 'react';
-import { toast } from 'react-toastify';
+import ManageVideosRow from './ManageVideosRow';
 
 const ManageVideos = () => {
     const url = 'https://infinite-island-65121.herokuapp.com/uploadedVideo';
@@ -14,29 +14,7 @@ const ManageVideos = () => {
             .then(res => res.json())
     )
     // console.log(data);
-    const deleteVideo = (id) => {
-        const url = `https://infinite-island-65121.herokuapp.com/uploadedVideo/${id}`;
-        fetch(url, {
-            method: 'DELETE',
-            headers: {
-                'content-type': 'application/json'
-            }
-        })
-
-            .then(res => res.json())
-            .then(result => {
-                alert('Are You sure???')
-                toast.success(`${id} is deleted`)
-                refetch()
-            })
-            .then(res => res.json())
-            .then(result => {
-                alert('Are You sure???')
-                toast.success(`${id} is deleted`)
-                refetch()
-            })
-    }
-
+    
     return (
         <div>
             <div className="overflow-x-auto w-full">
@@ -53,42 +31,13 @@ const ManageVideos = () => {
                     </thead>
                     <tbody>
                         {
-                            data?.map((detail, index) => <tr key={detail._id}>
-                                <td>
-                                    <span className='text-primary'>{index + 1}</span>
-                                </td>
-                                <td>
-                                    <div className="flex items-center space-x-3">
-                                        <div className="avatar">
-                                            <div className="mask mask-squircle w-12 h-12">
-                                                <img src={detail?.imgLink} alt="Avatar Tailwind CSS Component" />
-                                            </div>
-                                        </div>
-                                        <div>
-                                            <a href={detail?.videoLink} className='text-black' target="_blank" alt=''>{detail?.videoLink}</a>
-                                        </div>
-                                    </div>
-                                </td>
-
-                                <td>
-                                    <span className='text-black'>{detail?.uploader}</span>
-                                </td>
-                                <td>
-                                    <button
-                                        className="btn btn-success btn-xs"
-                                    >Final Upload</button>
-                                </td>
-                                <th>
-                                    <button
-                                        className="btn btn-error btn-xs"
-                                        onClick={() => deleteVideo(detail._id)}
-                                    >Delete</button>
-                                </th>
-                            </tr>)
+                            data?.map((detail, index) => <ManageVideosRow
+                            key={detail._id}
+                            detail={detail}
+                            index={index}
+                            refetch={refetch}
+                            ></ManageVideosRow>)
                         }
-                        <tr>
-
-                        </tr>
                     </tbody>
                     {/* <!-- foot --> */}
                     <tfoot>
